MS-9573 Remove javascript:void(0)
Removing the usage of `javascript:void(0)`.
This is an obsolete and incorrect practise for making elements, "do nothing".

EventHandlers are required to handle events correctly, and defaulting to `href="#"` or `href=""` is the correct approach in today's JavaScript world.

The `javascript:void(0)` results in CSP violations when the CSP is configured for web-applications.