MS-9555 Use Buttons instead of Links
To properly support `onClick` events, we should use a `<button>` element instead.

The pull request replaces the behavior of the Marionette ActionButton to use an actual HTML button element, and no longer an anchor.
Using Anchors causes pages to actually reload or follow the assigned target, which breaks the JavaScript functionality. Buttons will still be clickable, and can use the proper events as well.